test: 다리 구성 테스트 코드 작성
- 각각 건널 수 있는 경우와 건널 수 없는 경우의 테스트 코드 작성
